refactor: modify test suite to have separate workflows for different element types
