+++ cypress/cypress/e2e/D9-questions-case-study-workflow.cy.ts
@@ -1,0 +1,956 @@
+import messages from '../../../packages/i18n/messages/en'
+
+describe('Test creation and editing functionalities, validation, etc. for case study elements', function () {
+  beforeEach('Login the lecturer and load data fixture', function () {
+    cy.loginLecturer()
+    cy.fixture('D-questions.json').then((data) => {
+      this.data = data
+    })
+  })
+
+  // ! Case Study questions
+  // #region
+  it('Create the answer collection that will be used for the case study question tests', function () {
+    cy.get('[data-cy="resources"]').click()
+    cy.createAnswerCollection({
+      name: this.data.CS.collection,
+      accessCy: 'private',
+      access: messages.manage.resources.accessPRIVATE,
+      description: this.data.CS.collectionDescription,
+      entries: [...this.data.CS.items, ...this.data.CS.unselectedItems],
+    })
+    cy.createAnswerCollection({
+      name: this.data.CS.collectionEdited,
+      accessCy: 'restricted',
+      access: messages.manage.resources.accessRESTRICTED,
+      description: this.data.CS.collectionDescriptionEdited,
+      entries: [
+        ...this.data.CS.itemsEdited,
+        ...this.data.CS.unselectedItemsEdited,
+      ],
+    })
+  })
+
+  it('Create a Case Study question', function () {
+    cy.get('[data-cy="create-question"]').click()
+    cy.get('[data-cy="select-question-type"]')
+      .should('exist')
+      .contains(messages.shared.SC.typeLabel)
+    cy.get('[data-cy="select-question-type"]').click()
+    cy.get(
+      `[data-cy="select-question-type-${messages.shared.CASE_STUDY.typeLabel}"]`
+    ).click()
+    cy.get('[data-cy="select-question-type"]')
+      .should('exist')
+      .contains(messages.shared.CASE_STUDY.typeLabel)
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+
+    // enter question data
+    cy.get('[data-cy="insert-question-title"]').click().type(this.data.CS.title)
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="select-question-status"]').click()
+    cy.get(
+      `[data-cy="select-question-status-${messages.shared.READY.statusLabel}"]`
+    ).click()
+    cy.get('[data-cy="insert-question-text"]')
+      .realClick()
+      .type(this.data.CS.content)
+    cy.get('[data-cy="insert-question-explanation"]')
+      .realClick()
+      .type(this.data.CS.explanation)
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+
+    // select an answer collection
+    cy.get('[data-cy="select-answer-collection"]').contains(
+      messages.manage.questionForms.selectCollection
+    )
+    cy.get('[data-cy="select-answer-collection"]').click()
+    cy.get(
+      `[data-cy="select-answer-collection-${this.data.CS.collection}"]`
+    ).click()
+    cy.get('[data-cy="select-answer-collection"]').contains(
+      this.data.CS.collection
+    )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+
+    // select items for case study
+    cy.wrap(this.data.CS.items).each((item: string) => {
+      cy.get('[data-cy="choose-case-study-items"]').click()
+      cy.findByText(item).realClick()
+      cy.get('[data-cy="choose-case-study-items"]').contains(item)
+    })
+
+    // add new criteria, and remove one again
+    cy.wrap([...this.data.CS.criteria, this.data.CS.removedCriterion]).each(
+      (
+        criterion: {
+          name: string
+          min: number
+          max: number
+          step: number
+          unit?: string
+        },
+        ix
+      ) => {
+        cy.get('[data-cy="add-new-criterion"]').click()
+        cy.get(`[data-cy="criterion-${ix}-name"]`).click().type(criterion.name)
+        cy.get(`[data-cy="criterion-${ix}-min"]`)
+          .click()
+          .type(String(criterion.min))
+        cy.get(`[data-cy="criterion-${ix}-max"]`)
+          .click()
+          .type(String(criterion.max))
+        cy.get(`[data-cy="criterion-${ix}-step"]`)
+          .click()
+          .type(String(criterion.step))
+        if (criterion.unit) {
+          cy.get(`[data-cy="criterion-${ix}-unit"]`)
+            .click()
+            .type(criterion.unit)
+        }
+
+        cy.get(`[data-cy="criterion-${ix}-name"]`).should(
+          'have.value',
+          criterion.name
+        )
+        cy.get(`[data-cy="criterion-${ix}-min"]`).should(
+          'have.value',
+          String(criterion.min)
+        )
+        cy.get(`[data-cy="criterion-${ix}-max"]`).should(
+          'have.value',
+          String(criterion.max)
+        )
+        cy.get(`[data-cy="criterion-${ix}-step"]`).should(
+          'have.value',
+          String(criterion.step)
+        )
+        if (criterion.unit) {
+          cy.get(`[data-cy="criterion-${ix}-unit"]`).should(
+            'have.value',
+            criterion.unit
+          )
+        }
+      }
+    )
+    cy.get(
+      `[data-cy="remove-criterion-${this.data.CS.criteria.length}"]`
+    ).click()
+    cy.get(`[data-cy="criterion-${this.data.CS.criteria.length}-name"]`).should(
+      'not.exist'
+    )
+    cy.get(`[data-cy="criterion-${this.data.CS.criteria.length}-min"]`).should(
+      'not.exist'
+    )
+    cy.get(`[data-cy="criterion-${this.data.CS.criteria.length}-max"]`).should(
+      'not.exist'
+    )
+    cy.get(`[data-cy="criterion-${this.data.CS.criteria.length}-step"]`).should(
+      'not.exist'
+    )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+
+    cy.wrap([...this.data.CS.cases, this.data.CS.removedCase]).each(
+      (caseItem: { title: string; description: string }, ix) => {
+        cy.get('[data-cy="add-new-case"]').click()
+        cy.get(`[data-cy="case-title-${ix}"]`).click().type(caseItem.title)
+        cy.get(`[data-cy="case-description-${ix}"]`)
+          .realClick()
+          .type(caseItem.description)
+
+        cy.get(`[data-cy="case-title-${ix}"]`).should(
+          'have.value',
+          caseItem.title
+        )
+        cy.get(`[data-cy="case-description-${ix}"]`).contains(
+          caseItem.description
+        )
+        cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+      }
+    )
+    cy.get(`[data-cy="delete-case-${this.data.CS.cases.length}"]`).click()
+    cy.get(`[data-cy="cancel-delete-case"]`).click()
+    cy.get(`[data-cy="delete-case-${this.data.CS.cases.length}"]`).click()
+    cy.get(`[data-cy="confirm-delete-case"]`).click()
+    cy.get(`[data-cy="case-title-${this.data.CS.cases.length}"]`).should(
+      'not.exist'
+    )
+    cy.get(`[data-cy="case-description-${this.data.CS.cases.length}"]`).should(
+      'not.exist'
+    )
+
+    // test that enabling sample solution works correctly
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="save-new-question"]').click()
+    cy.wait(500)
+
+    cy.get(`[data-cy="element-item-${this.data.CS.title}"]`).contains(
+      this.data.CS.content
+    )
+    cy.get(`[data-cy="element-item-${this.data.CS.title}"]`).contains(
+      this.data.CS.title
+    )
+    cy.get(`[data-cy="element-item-${this.data.CS.title}"]`).contains(
+      messages.shared.READY.statusLabel
+    )
+  })
+
+  it('Verify that the correct content has been saved', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.title}"]`).click()
+    cy.get('[data-cy="insert-question-title"]').should(
+      'have.value',
+      this.data.CS.title
+    )
+    cy.get('[data-cy="select-question-status"]').contains(
+      messages.shared.READY.statusLabel
+    )
+    cy.get('[data-cy="insert-question-text"]').contains(this.data.CS.content)
+    cy.get('[data-cy="insert-question-explanation"]').contains(
+      this.data.CS.explanation
+    )
+
+    cy.get('[data-cy="select-answer-collection"]').contains(
+      this.data.CS.collection
+    )
+    cy.wrap(this.data.CS.items).each((item: string) => {
+      cy.get('[data-cy="choose-case-study-items"]').contains(item)
+    })
+
+    cy.wrap(this.data.CS.criteria).each(
+      (
+        criterion: {
+          name: string
+          min: number
+          max: number
+          step: number
+          unit?: string
+        },
+        ix
+      ) => {
+        cy.get(`[data-cy="criterion-${ix}-name"]`).should(
+          'have.value',
+          criterion.name
+        )
+        cy.get(`[data-cy="criterion-${ix}-min"]`).should(
+          'have.value',
+          String(criterion.min)
+        )
+        cy.get(`[data-cy="criterion-${ix}-max"]`).should(
+          'have.value',
+          String(criterion.max)
+        )
+        cy.get(`[data-cy="criterion-${ix}-step"]`).should(
+          'have.value',
+          String(criterion.step)
+        )
+        if (criterion.unit) {
+          cy.get(`[data-cy="criterion-${ix}-unit"]`).should(
+            'have.value',
+            criterion.unit
+          )
+        }
+      }
+    )
+
+    cy.wrap(this.data.CS.cases).each(
+      (caseItem: { title: string; description: string }, ix) => {
+        cy.get(`[data-cy="case-title-${ix}"]`).should(
+          'have.value',
+          caseItem.title
+        )
+        cy.get(`[data-cy="case-description-${ix}"]`).contains(
+          caseItem.description
+        )
+      }
+    )
+
+    cy.get('[data-cy="close-element-modal"]').click()
+  })
+
+  it('Verify that creation was successful', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.title}"]`).click()
+    cy.get('[data-cy="insert-question-title"]').should(
+      'have.value',
+      this.data.CS.title
+    )
+
+    cy.get('[data-cy="student-element-preview"]')
+      .findByText(messages.shared.generic.instructions)
+      .should('exist')
+    cy.get('[data-cy="student-element-preview"]') // instructions should be visible in preview
+      .findByText(this.data.CS.content)
+      .should('exist')
+
+    // check if case information is visible
+    cy.get('[data-cy="case-0-title"]').contains(this.data.CS.cases[0].title)
+    cy.get('[data-cy="case-0-description"]').contains(
+      this.data.CS.cases[0].description
+    )
+    cy.get('[data-cy="case-1-title"]').contains(this.data.CS.cases[1].title)
+    cy.get('[data-cy="case-1-description"]').contains(
+      this.data.CS.cases[1].description
+    )
+
+    // check that sliders are initilized correctly and that values changes persist
+    const steps = 78
+    const midValue =
+      this.data.CS.criteria[0].min +
+      (this.data.CS.criteria[0].max - this.data.CS.criteria[0].min) / 2
+    const slidedValue = midValue + steps * this.data.CS.criteria[0].step
+    cy.get('[data-cy="cs-slider-nr-value-0-0-0-0"]').should('have.value', '')
+    cy.get('[data-cy="cs-slider-0-0-0-0"]')
+      .click()
+      .type('{rightarrow}{leftarrow}')
+    cy.get('[data-cy="cs-slider-nr-value-0-0-0-0"]').should(
+      'have.value',
+      String(midValue)
+    )
+    cy.get('[data-cy="cs-slider-0-0-0-0"]')
+      .click()
+      .type('{rightarrow}'.repeat(steps))
+    cy.get('[data-cy="cs-slider-nr-value-0-0-0-0"]').should(
+      'have.value',
+      String(slidedValue)
+    )
+
+    // check that moving a slider all the way to one end works to be expected
+    cy.get('[data-cy="cs-slider-nr-value-0-0-0-1"]').should('have.value', '')
+    cy.get('[data-cy="cs-slider-0-0-0-1"]')
+      .click()
+      .type('{leftarrow}'.repeat(260))
+    cy.get('[data-cy="cs-slider-nr-value-0-0-0-1"]').should(
+      'have.value',
+      String(this.data.CS.criteria[1].min)
+    )
+    cy.get('[data-cy="cs-slider-0-0-0-1"]')
+      .click()
+      .type('{rightarrow}'.repeat(600))
+    cy.get('[data-cy="cs-slider-nr-value-0-0-0-1"]').should(
+      'have.value',
+      String(this.data.CS.criteria[1].max)
+    )
+
+    // check that sliders are shown for all response items
+    for (let caseIx = 0; caseIx < this.data.CS.cases.length; caseIx++) {
+      for (
+        let criterionIx = 0;
+        criterionIx < this.data.CS.criteria.length;
+        criterionIx++
+      ) {
+        for (let itemIx = 0; itemIx < this.data.CS.items.length; itemIx++) {
+          cy.get(
+            `[data-cy="cs-slider-nr-value-0-${caseIx}-${itemIx}-${criterionIx}"]`
+          ).should('exist')
+        }
+      }
+    }
+  })
+
+  it('Verify that the deletion of answer collection entries is limited, editing is unaffected', function () {
+    cy.get('[data-cy="resources"]').click()
+    cy.get(`[data-cy="answer-collection-${this.data.CS.collection}"]`).click()
+
+    cy.wrap(this.data.CS.items).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should('be.disabled')
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.wrap(this.data.CS.unselectedItems).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.get("[data-cy='close-answer-collection-edit-modal']").click()
+  })
+
+  it('Verify that the answer collection used in the case study can no longer be deleted', function () {
+    cy.get('[data-cy="resources"]').click()
+    cy.get(`[data-cy="answer-collection-${this.data.CS.collection}"]`).click()
+    cy.findByText(messages.manage.resources.answerOptionUsed).should('exist')
+    cy.get('[data-cy="delete-answer-collection"]').should('be.disabled')
+  })
+
+  it('Add a sample solution to the case study question', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.title}"]`).click()
+    cy.get('[data-cy="insert-question-title"]').should(
+      'have.value',
+      this.data.CS.title
+    )
+
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled') // correct answers for all criteria & items are required
+    cy.caseStudyLoop({
+      object: this.data.CS.solutions,
+      callback: ({ caseIx, itemIx, criterionIx, innerValue }) => {
+        const value = innerValue as { lower: number; upper: number }
+
+        cy.get('[data-cy="save-new-question"]').should('be.disabled')
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-lower"]`
+        )
+          .click()
+          .type(String(value.lower))
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-upper"]`
+        )
+          .click()
+          .type(String(value.upper))
+
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-lower"]`
+        ).should('have.value', String(value.lower))
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-upper"]`
+        ).should('have.value', String(value.upper))
+      },
+    })
+    cy.get('[data-cy="save-new-question"]').click()
+    cy.wait(500)
+  })
+
+  it('Verify that the sample solution has been stored correctly for the modified case study question', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.title}"]`).click()
+    cy.get('[data-cy="insert-question-title"]').should(
+      'have.value',
+      this.data.CS.title
+    )
+
+    cy.caseStudyLoop({
+      object: this.data.CS.solutions,
+      callback: ({ caseIx, itemIx, criterionIx, innerValue }) => {
+        const value = innerValue as { lower: number; upper: number }
+
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-lower"]`
+        ).should('have.value', String(value.lower))
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-upper"]`
+        ).should('have.value', String(value.upper))
+      },
+    })
+
+    cy.get('[data-cy="close-element-modal"]').click()
+  })
+
+  it('Verify that the case study validation logic covers all required cases and block submission of invalid element edit modals', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.title}"]`).click()
+
+    // missing question title -> invalid
+    cy.get('[data-cy="insert-question-title"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="insert-question-title"]').click().type(this.data.CS.title)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // missing question content -> invalid
+    cy.get('[data-cy="insert-question-text"]').realClick().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="insert-question-text"]')
+      .realClick()
+      .type(this.data.CS.content)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // missing question explanation -> valid
+    cy.get('[data-cy="insert-question-explanation"]').realClick().clear()
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="insert-question-explanation"]')
+      .realClick()
+      .type(this.data.CS.explanation)
+
+    // criteria name, min, max, step required -> invalid (if removed)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="criterion-0-name"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="criterion-0-name"]')
+      .click()
+      .type(this.data.CS.criteria[0].name)
+
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="criterion-0-min"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="criterion-0-min"]')
+      .click()
+      .type(String(this.data.CS.criteria[0].min))
+
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="criterion-0-max"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="criterion-0-max"]')
+      .click()
+      .type(String(this.data.CS.criteria[0].max))
+
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="criterion-0-step"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="criterion-0-step"]')
+      .click()
+      .type(String(this.data.CS.criteria[0].step))
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // criterion min <= max required & max - min >= 2 * step -> otherwise invalid
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="criterion-0-min"]')
+      .click()
+      .clear()
+      .type(
+        String(this.data.CS.criteria[0].max + this.data.CS.criteria[0].step + 1)
+      )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="criterion-0-min"]')
+      .click()
+      .clear()
+      .type(
+        String(
+          this.data.CS.criteria[0].max - 2 * this.data.CS.criteria[0].step + 1
+        )
+      )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="criterion-0-min"]')
+      .click()
+      .clear()
+      .type(
+        String(
+          this.data.CS.criteria[0].max - 2 * this.data.CS.criteria[0].step - 1
+        )
+      )
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="criterion-0-min"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].min))
+    cy.get('[data-cy="configure-sample-solution"]').click()
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // solutions: lower and upper bound required
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .type(this.data.CS.solutions[1][3][0].lower)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    cy.get('[data-cy="case-solution-1-3-0-upper"]').click().clear()
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-upper"]')
+      .click()
+      .type(this.data.CS.solutions[1][3][0].upper)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // solutions: min <= max required
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(this.data.CS.solutions[1][3][0].upper + 1)
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(this.data.CS.solutions[1][3][0].lower)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // solutions: min and max lie within the bounds of the criterion
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].min - 1))
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].min))
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].min + 1))
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(this.data.CS.solutions[1][3][0].lower)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-upper"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].max + 1))
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-upper"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].max))
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-upper"]')
+      .click()
+      .clear()
+      .type(String(this.data.CS.criteria[0].max - 1))
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-upper"]')
+      .click()
+      .clear()
+      .type(this.data.CS.solutions[1][3][0].upper)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+
+    // solutions: max - min >= step size
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(
+        String(
+          this.data.CS.solutions[1][3][0].upper -
+            this.data.CS.criteria[0].step +
+            1
+        )
+      )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(
+        String(
+          this.data.CS.solutions[1][3][0].upper - this.data.CS.criteria[0].step
+        )
+      )
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(
+        String(
+          this.data.CS.solutions[1][3][0].upper -
+            this.data.CS.criteria[0].step -
+            1
+        )
+      )
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+    cy.get('[data-cy="case-solution-1-3-0-lower"]')
+      .click()
+      .clear()
+      .type(this.data.CS.solutions[1][3][0].lower)
+    cy.get('[data-cy="save-new-question"]').should('not.be.disabled')
+  })
+
+  it('Edit the case study question and change the answer collection (including new sample solutions)', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.title}"]`).click()
+    cy.get('[data-cy="insert-question-title"]')
+      .click()
+      .clear()
+      .type(this.data.CS.titleEdited)
+    cy.get('[data-cy="insert-question-text"]')
+      .realClick()
+      .clear()
+      .type(this.data.CS.contentEdited)
+
+    cy.get('[data-cy="select-answer-collection"]').click()
+    cy.get(
+      `[data-cy="select-answer-collection-${this.data.CS.collectionEdited}"]`
+    ).click()
+    cy.get('[data-cy="cancel-change-collection"]').click()
+    cy.get('[data-cy="select-answer-collection"]').click()
+    cy.get(
+      `[data-cy="select-answer-collection-${this.data.CS.collectionEdited}"]`
+    ).click()
+    cy.get('[data-cy="confirm-change-collection"]').click()
+    cy.get('[data-cy="select-answer-collection"]').contains(
+      this.data.CS.collectionEdited
+    )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled') // answer options are cleared on collection change
+
+    // select items for case study
+    cy.wrap(this.data.CS.itemsEdited).each((item: string) => {
+      cy.get('[data-cy="choose-case-study-items"]').click()
+      cy.findByText(item).realClick()
+      cy.get('[data-cy="choose-case-study-items"]').contains(item)
+    })
+
+    // clear all fields, enter new criteria
+    cy.wrap(this.data.CS.criteriaEdited).each(
+      (
+        criterion: {
+          name: string
+          min: number
+          max: number
+          step: number
+          unit?: string
+        },
+        ix
+      ) => {
+        cy.get(`[data-cy="criterion-${ix}-name"]`)
+          .click()
+          .clear()
+          .type(criterion.name)
+        cy.get(`[data-cy="criterion-${ix}-min"]`)
+          .click()
+          .clear()
+          .type(String(criterion.min))
+        cy.get(`[data-cy="criterion-${ix}-max"]`)
+          .click()
+          .clear()
+          .type(String(criterion.max))
+        cy.get(`[data-cy="criterion-${ix}-step"]`)
+          .click()
+          .clear()
+          .type(String(criterion.step))
+        cy.get(`[data-cy="criterion-${ix}-unit"]`).click().clear()
+
+        if (criterion.unit) {
+          cy.get(`[data-cy="criterion-${ix}-unit"]`)
+            .click()
+            .type(criterion.unit)
+        }
+
+        cy.get(`[data-cy="criterion-${ix}-name"]`).should(
+          'have.value',
+          criterion.name
+        )
+        cy.get(`[data-cy="criterion-${ix}-min"]`).should(
+          'have.value',
+          String(criterion.min)
+        )
+        cy.get(`[data-cy="criterion-${ix}-max"]`).should(
+          'have.value',
+          String(criterion.max)
+        )
+        cy.get(`[data-cy="criterion-${ix}-step"]`).should(
+          'have.value',
+          String(criterion.step)
+        )
+        if (criterion.unit) {
+          cy.get(`[data-cy="criterion-${ix}-unit"]`).should(
+            'have.value',
+            criterion.unit
+          )
+        }
+      }
+    )
+
+    // remove all existing cases
+    for (let i = 0; i < this.data.CS.cases.length; i++) {
+      cy.get(`[data-cy="delete-case-0"]`).click()
+      cy.get(`[data-cy="confirm-delete-case"]`).click()
+    }
+    cy.get('[data-cy="save-new-question"]').should('be.disabled') // cases required
+
+    // add new cases
+    cy.wrap(this.data.CS.casesEdited).each(
+      (caseItem: { title: string; description: string }, ix) => {
+        cy.get('[data-cy="add-new-case"]').click()
+        cy.get(`[data-cy="case-title-${ix}"]`).click().type(caseItem.title)
+        cy.get(`[data-cy="case-description-${ix}"]`)
+          .realClick()
+          .type(caseItem.description)
+
+        cy.get(`[data-cy="case-title-${ix}"]`).should(
+          'have.value',
+          caseItem.title
+        )
+        cy.get(`[data-cy="case-description-${ix}"]`).contains(
+          caseItem.description
+        )
+      }
+    )
+    cy.get('[data-cy="save-new-question"]').should('be.disabled') // solution required
+
+    // add new sample solutions
+    cy.caseStudyLoop({
+      object: this.data.CS.solutionsEdited,
+      callback: ({ caseIx, itemIx, criterionIx, innerValue }) => {
+        const value = innerValue as { lower: number; upper: number }
+
+        cy.get('[data-cy="save-new-question"]').should('be.disabled')
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-lower"]`
+        )
+          .click()
+          .type(String(value.lower))
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-upper"]`
+        )
+          .click()
+          .type(String(value.upper))
+      },
+    })
+    cy.get('[data-cy="save-new-question"]').click()
+  })
+
+  it('Verify that all changes to the case study question have been saved correctly', function () {
+    cy.get(`[data-cy="edit-question-${this.data.CS.titleEdited}"]`).click()
+    cy.get('[data-cy="insert-question-title"]').should(
+      'have.value',
+      this.data.CS.titleEdited
+    )
+    cy.get('[data-cy="insert-question-text"]')
+      .realClick()
+      .contains(this.data.CS.contentEdited)
+
+    cy.get('[data-cy="select-answer-collection"]').contains(
+      this.data.CS.collectionEdited
+    )
+    cy.wrap(this.data.CS.itemsEdited).each((item: string) => {
+      cy.get('[data-cy="choose-case-study-items"]').contains(item)
+    })
+
+    cy.wrap(this.data.CS.criteriaEdited).each(
+      (
+        criterion: {
+          name: string
+          min: number
+          max: number
+          step: number
+          unit?: string
+        },
+        ix
+      ) => {
+        cy.get(`[data-cy="criterion-${ix}-name"]`).should(
+          'have.value',
+          criterion.name
+        )
+        cy.get(`[data-cy="criterion-${ix}-min"]`).should(
+          'have.value',
+          String(criterion.min)
+        )
+        cy.get(`[data-cy="criterion-${ix}-max"]`).should(
+          'have.value',
+          String(criterion.max)
+        )
+        cy.get(`[data-cy="criterion-${ix}-step"]`).should(
+          'have.value',
+          String(criterion.step)
+        )
+        if (criterion.unit) {
+          cy.get(`[data-cy="criterion-${ix}-unit"]`).should(
+            'have.value',
+            criterion.unit
+          )
+        }
+      }
+    )
+
+    cy.wrap(this.data.CS.casesEdited).each(
+      (caseItem: { title: string; description: string }, ix) => {
+        cy.get(`[data-cy="case-title-${ix}"]`).should(
+          'have.value',
+          caseItem.title
+        )
+        cy.get(`[data-cy="case-description-${ix}"]`).contains(
+          caseItem.description
+        )
+      }
+    )
+
+    cy.caseStudyLoop({
+      object: this.data.CS.solutionsEdited,
+      callback: ({ caseIx, itemIx, criterionIx, innerValue }) => {
+        const value = innerValue as { lower: number; upper: number }
+
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-lower"]`
+        ).should('have.value', String(value.lower))
+        cy.get(
+          `[data-cy="case-solution-${caseIx}-${itemIx}-${criterionIx}-upper"]`
+        ).should('have.value', String(value.upper))
+      },
+    })
+
+    cy.get('[data-cy="close-element-modal"]').click()
+  })
+
+  it('Verify that all elements of the previously used answer collection and the collection itself can be deleted again', function () {
+    cy.get('[data-cy="resources"]').click()
+
+    cy.get(`[data-cy="answer-collection-${this.data.CS.collection}"]`).click()
+    cy.wrap(this.data.CS.items).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.wrap(this.data.CS.unselectedItems).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.findByText(messages.manage.resources.answerOptionUsed).should(
+      'not.exist'
+    )
+    cy.get('[data-cy="delete-answer-collection"]').should('not.be.disabled')
+    cy.get("[data-cy='close-answer-collection-edit-modal']").click()
+
+    cy.get(
+      `[data-cy="answer-collection-${this.data.CS.collectionEdited}"]`
+    ).click()
+    cy.wrap(this.data.CS.itemsEdited).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should('be.disabled')
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.wrap(this.data.CS.unselectedItemsEdited).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.findByText(messages.manage.resources.answerOptionUsed).should('exist')
+    cy.get('[data-cy="delete-answer-collection"]').should('be.disabled')
+    cy.get("[data-cy='close-answer-collection-edit-modal']").click()
+  })
+  // #endregion
+
+  // ! Cleanup
+  // #region
+  it('Cleanup: Delete case study question', function () {
+    cy.deleteElement({ elementName: this.data.CS.titleEdited })
+  })
+
+  it('Verify that after the deletion of the linked questions, all solution options can be deleted again', function () {
+    cy.get('[data-cy="resources"]').click()
+
+    cy.get(`[data-cy="answer-collection-${this.data.CS.collection}"]`).click()
+    cy.wrap(this.data.CS.items).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.wrap(this.data.CS.unselectedItems).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.get("[data-cy='close-answer-collection-edit-modal']").click()
+
+    cy.get(
+      `[data-cy="answer-collection-${this.data.CS.collectionEdited}"]`
+    ).click()
+    cy.wrap(this.data.CS.itemsEdited).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.wrap(this.data.CS.unselectedItemsEdited).each((sol: string) => {
+      cy.get(`[data-cy="delete-answer-option-${sol}"]`).should(
+        'not.be.disabled'
+      )
+      cy.get(`[data-cy="edit-answer-option-${sol}"]`).should('not.be.disabled')
+    })
+    cy.get("[data-cy='close-answer-collection-edit-modal']").click()
+  })
+
+  it('Cleanup: Delete all created answer collections', function () {
+    cy.get('[data-cy="resources"]').click()
+    cy.deleteAnswerCollection({ collectionName: this.data.CS.collection })
+    cy.deleteAnswerCollection({ collectionName: this.data.CS.collectionEdited })
+  })
+  // #endregion
+})
